requirements.txt
Flask
flask-cors
PyMuPDF
python-docx
scikit-learn==1.3.2
numpy==1.26.4
openai
lxml==4.9.3
scipy==1.11.4